[Draft] Split Draft Snap Gui Commands from DraftTools
bugfix


bugfix2
