+++ src/Mod/Draft/DraftTools.py
@@ -5394,232 +5394,6 @@ class Draft_Arc_3Points:
 
 
 #---------------------------------------------------------------------------
-# Snap tools
-#---------------------------------------------------------------------------
-
-def get_snap_toolbar():
-    from PySide import QtGui
-    mw = FreeCADGui.getMainWindow()
-    if mw:
-        toolbar = mw.findChild(QtGui.QToolBar,"Draft Snap")
-        if toolbar:
-            return toolbar
-    return None
-
-def sync_snap_toolbar_button(button, status):
-    snap_toolbar = get_snap_toolbar()
-    if snap_toolbar:
-        for a in snap_toolbar.actions():
-            if a.objectName() == button:
-                a.setChecked(status)
-                if a.isChecked():
-                    a.setToolTip(a.toolTip().replace("OFF","ON"))
-                else:
-                    a.setToolTip(a.toolTip().replace("ON","OFF"))
-
-def sync_snap_statusbar_button(button, status):
-    mw = FreeCADGui.getMainWindow()
-    if mw:
-        sb = mw.statusBar()
-        if sb:
-            ortho_button = sb.findChild(QtGui.QAction,button)
-            if ortho_button:
-                ortho_button.setChecked(status)
-
-class Draft_Snap_Lock():
-    def GetResources(self):
-        return {'Pixmap'  : 'Snap_Lock',
-                'Accel' : "Shift+S",
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Lock", "Toggle On/Off"),
-                'ToolTip' : QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Lock", "Activates/deactivates all snap tools at once")}
-    def Activated(self):
-        if hasattr(FreeCADGui,"Snapper"):
-            # toggle the corresponding snap_index in Preferences/Mod/Draft/snapModes
-            status = FreeCADGui.Snapper.toggle_snap('Lock')
-            # change interface consistently
-            sync_snap_toolbar_button("Draft_Snap_Lock"+"_Button", status)
-            snap_toolbar = get_snap_toolbar()
-            if snap_toolbar:
-                snap_toolbar.actions()[0].setChecked(status)
-                for a in snap_toolbar.actions()[1:]:
-                    a.setEnabled(status)
-
-class Draft_Snap_Near():
-    def GetResources(self):
-        return {'Pixmap'  : 'Snap_Near',
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Near", "Nearest"),
-                'ToolTip' : QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Near", "Snaps to nearest point on edges")}
-    def Activated(self):
-        if hasattr(FreeCADGui,"Snapper"):
-            # toggle the corresponding snap_index in Preferences/Mod/Draft/snapModes
-            status = FreeCADGui.Snapper.toggle_snap('Near')
-            # change interface consistently
-            sync_snap_toolbar_button("Draft_Snap_Near"+"_Button", status)
-
-
-class Draft_Snap_Midpoint():
-    def GetResources(self):
-        return {'Pixmap'  : 'Snap_Midpoint',
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Midpoint", "Midpoint"),
-                'ToolTip' : QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Midpoint", "Snaps to midpoints of edges")}
-    def Activated(self):
-        if hasattr(FreeCADGui,"Snapper"):
-            # toggle the corresponding snap_index in Preferences/Mod/Draft/snapModes
-            status = FreeCADGui.Snapper.toggle_snap('Midpoint')
-            # change interface consistently
-            sync_snap_toolbar_button("Draft_Snap_Midpoint"+"_Button", status)
-
-class Draft_Snap_Perpendicular():
-    def GetResources(self):
-        return {'Pixmap'  : 'Snap_Perpendicular',
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Perpendicular", "Perpendicular"),
-                'ToolTip' : QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Perpendicular", "Snaps to perpendicular points on edges")}
-    def Activated(self):
-        if hasattr(FreeCADGui,"Snapper"):
-            # toggle the corresponding snap_index in Preferences/Mod/Draft/snapModes
-            status = FreeCADGui.Snapper.toggle_snap('Perpendicular')
-            # change interface consistently
-            sync_snap_toolbar_button("Draft_Snap_Perpendicular"+"_Button", status)
-
-class Draft_Snap_Grid():
-    def GetResources(self):
-        return {'Pixmap'  : 'Snap_Grid',
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Grid", "Grid"),
-                'ToolTip' : QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Grid", "Snaps to grid points")}
-    def Activated(self):
-        if hasattr(FreeCADGui,"Snapper"):
-            # toggle the corresponding snap_index in Preferences/Mod/Draft/snapModes
-            status = FreeCADGui.Snapper.toggle_snap('Grid')
-            # change interface consistently
-            sync_snap_toolbar_button("Draft_Snap_Grid"+"_Button", status)
-
-class Draft_Snap_Intersection():
-    def GetResources(self):
-        return {'Pixmap'  : 'Snap_Intersection',
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Intersection", "Intersection"),
-                'ToolTip' : QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Intersection", "Snaps to edges intersections")}
-    def Activated(self):
-        if hasattr(FreeCADGui,"Snapper"):
-            # toggle the corresponding snap_index in Preferences/Mod/Draft/snapModes
-            status = FreeCADGui.Snapper.toggle_snap('Intersection')
-            # change interface consistently
-            sync_snap_toolbar_button("Draft_Snap_Intersection"+"_Button", status)
-
-class Draft_Snap_Parallel():
-    def GetResources(self):
-        return {'Pixmap'  : 'Snap_Parallel',
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Parallel", "Parallel"),
-                'ToolTip' : QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Parallel", "Snaps to parallel directions of edges")}
-    def Activated(self):
-        if hasattr(FreeCADGui,"Snapper"):
-            # toggle the corresponding snap_index in Preferences/Mod/Draft/snapModes
-            status = FreeCADGui.Snapper.toggle_snap('Parallel')
-            # change interface consistently
-            sync_snap_toolbar_button("Draft_Snap_Parallel"+"_Button", status)
-
-class Draft_Snap_Endpoint():
-    def GetResources(self):
-        return {'Pixmap'  : 'Snap_Endpoint',
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Endpoint", "Endpoint"),
-                'ToolTip' : QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Endpoint", "Snaps to endpoints of edges")}
-    def Activated(self):
-        if hasattr(FreeCADGui,"Snapper"):
-            # toggle the corresponding snap_index in Preferences/Mod/Draft/snapModes
-            status = FreeCADGui.Snapper.toggle_snap('Endpoint')
-            # change interface consistently
-            sync_snap_toolbar_button("Draft_Snap_Endpoint"+"_Button", status)
-
-class Draft_Snap_Angle():
-    def GetResources(self):
-        return {'Pixmap'  : 'Snap_Angle',
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Angle", "Angles"),
-                'ToolTip' : QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Angle", "Snaps to 45 and 90 degrees points on arcs and circles")}
-    def Activated(self):
-        if hasattr(FreeCADGui,"Snapper"):
-            # toggle the corresponding snap_index in Preferences/Mod/Draft/snapModes
-            status = FreeCADGui.Snapper.toggle_snap('Angle')
-            # change interface consistently
-            sync_snap_toolbar_button("Draft_Snap_Angle"+"_Button", status)
-
-class Draft_Snap_Center():
-    def GetResources(self):
-        return {'Pixmap'  : 'Snap_Center',
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Center", "Center"),
-                'ToolTip' : QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Center", "Snaps to center of circles and arcs")}
-    def Activated(self):
-        if hasattr(FreeCADGui,"Snapper"):
-            # toggle the corresponding snap_index in Preferences/Mod/Draft/snapModes
-            status = FreeCADGui.Snapper.toggle_snap('Center')
-            # change interface consistently
-            sync_snap_toolbar_button("Draft_Snap_Center"+"_Button", status)
-
-class Draft_Snap_Extension():
-    def GetResources(self):
-        return {'Pixmap'  : 'Snap_Extension',
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Extension", "Extension"),
-                'ToolTip' : QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Extension", "Snaps to extension of edges")}
-    def Activated(self):
-        if hasattr(FreeCADGui,"Snapper"):
-            # toggle the corresponding snap_index in Preferences/Mod/Draft/snapModes
-            status = FreeCADGui.Snapper.toggle_snap('Extension')
-            # change interface consistently
-            sync_snap_toolbar_button("Draft_Snap_Extension"+"_Button", status)
-
-class Draft_Snap_Ortho():
-    """
-    Toggle Snap Ortho (snap_index = 10)
-    """
-    def GetResources(self):
-        return {'Pixmap'  : 'Snap_Ortho',
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Ortho", "Ortho"),
-                'ToolTip' : QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Ortho", "Snaps to orthogonal and 45 degrees directions")}
-    
-    def Activated(self):
-        if hasattr(FreeCADGui,"Snapper"):
-            # toggle the corresponding snap_index in Preferences/Mod/Draft/snapModes
-            status = FreeCADGui.Snapper.toggle_snap('Ortho')
-            # change interface consistently
-            sync_snap_toolbar_button("Draft_Snap_Ortho"+"_Button", status)
-            sync_snap_statusbar_button("OrthoButton", status)
-
-class Draft_Snap_Special():
-    def GetResources(self):
-        return {'Pixmap'  : 'Snap_Special',
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Special", "Special"),
-                'ToolTip' : QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Special", "Snaps to special locations of objects")}
-    def Activated(self):
-        if hasattr(FreeCADGui,"Snapper"):
-            # toggle the corresponding snap_index in Preferences/Mod/Draft/snapModes
-            status = FreeCADGui.Snapper.toggle_snap('Special')
-            # change interface consistently
-            sync_snap_toolbar_button("Draft_Snap_Special"+"_Button", status)
-
-class Draft_Snap_Dimensions():
-    def GetResources(self):
-        return {'Pixmap'  : 'Snap_Dimensions',
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Dimensions", "Dimensions"),
-                'ToolTip' : QtCore.QT_TRANSLATE_NOOP("Draft_Snap_Dimensions", "Shows temporary dimensions when snapping to Arch objects")}
-    def Activated(self):
-        if hasattr(FreeCADGui,"Snapper"):
-            # toggle the corresponding snap_index in Preferences/Mod/Draft/snapModes
-            status = FreeCADGui.Snapper.toggle_snap('Dimensions')
-            # change interface consistently
-            sync_snap_toolbar_button("Draft_Snap_Dimensions"+"_Button", status)
-            sync_snap_statusbar_button("DimButton", status)
-
-class Draft_Snap_WorkingPlane():
-    def GetResources(self):
-        return {'Pixmap'  : 'Snap_WorkingPlane',
-                'MenuText': QtCore.QT_TRANSLATE_NOOP("Draft_Snap_WorkingPlane", "Working Plane"),
-                'ToolTip' : QtCore.QT_TRANSLATE_NOOP("Draft_Snap_WorkingPlane", "Restricts the snapped point to the current working plane")}
-    def Activated(self):
-        if hasattr(FreeCADGui,"Snapper"):
-            # toggle the corresponding snap_index in Preferences/Mod/Draft/snapModes
-            status = FreeCADGui.Snapper.toggle_snap('WorkingPlane')
-            # change interface consistently
-            sync_snap_toolbar_button("Draft_Snap_WorkingPlane"+"_Button", status)
-
-#---------------------------------------------------------------------------
 # Adds the icons & commands to the FreeCAD command manager, and sets defaults
 #---------------------------------------------------------------------------
 
@@ -5704,23 +5478,6 @@ FreeCADGui.addCommand('Draft_ToggleGrid',ToggleGrid())
 FreeCADGui.addCommand('Draft_FlipDimension',Draft_FlipDimension())
 FreeCADGui.addCommand('Draft_AutoGroup',SetAutoGroup())
 FreeCADGui.addCommand('Draft_AddConstruction',Draft_AddConstruction())
-
-# snap commands
-FreeCADGui.addCommand('Draft_Snap_Lock',Draft_Snap_Lock())
-FreeCADGui.addCommand('Draft_Snap_Midpoint',Draft_Snap_Midpoint())
-FreeCADGui.addCommand('Draft_Snap_Perpendicular',Draft_Snap_Perpendicular())
-FreeCADGui.addCommand('Draft_Snap_Grid',Draft_Snap_Grid())
-FreeCADGui.addCommand('Draft_Snap_Intersection',Draft_Snap_Intersection())
-FreeCADGui.addCommand('Draft_Snap_Parallel',Draft_Snap_Parallel())
-FreeCADGui.addCommand('Draft_Snap_Endpoint',Draft_Snap_Endpoint())
-FreeCADGui.addCommand('Draft_Snap_Angle',Draft_Snap_Angle())
-FreeCADGui.addCommand('Draft_Snap_Center',Draft_Snap_Center())
-FreeCADGui.addCommand('Draft_Snap_Extension',Draft_Snap_Extension())
-FreeCADGui.addCommand('Draft_Snap_Near',Draft_Snap_Near())
-FreeCADGui.addCommand('Draft_Snap_Ortho',Draft_Snap_Ortho())
-FreeCADGui.addCommand('Draft_Snap_Special',Draft_Snap_Special())
-FreeCADGui.addCommand('Draft_Snap_Dimensions',Draft_Snap_Dimensions())
-FreeCADGui.addCommand('Draft_Snap_WorkingPlane',Draft_Snap_WorkingPlane())
 
 # a global place to look for active draft Command
 FreeCAD.activeDraftCommand = None
