+++ src/Mod/Draft/InitGui.py
@@ -84,6 +84,7 @@ class DraftWorkbench(FreeCADGui.Workbench):
             from draftguitools import gui_polararray
             from draftguitools import gui_orthoarray
             from draftguitools import gui_arrays
+            from draftguitools import gui_snap           
             FreeCADGui.addLanguagePath(":/translations")
             FreeCADGui.addIconPath(":/icons")
         except Exception as exc:
